+++ app/components/DesktopActionsButtons/index.js
@@ -79,7 +79,7 @@ class DesktopActionsButtons extends React.Component {
     if (
       e.which === 27 &&
       this.state.showTransferMenu === true &&
-      this.props.interaction.direction !== 'outbound' &&
+      this.props.interaction.direction !== 'agent-initiated' &&
       this.props.interaction.status !== 'wrapup'
     ) {
       this.toggleTransferMenu();
@@ -118,7 +118,7 @@ class DesktopActionsButtons extends React.Component {
           <WrapUpToggle interaction={this.props.interaction} type="" />
         )}
         {this.props.interaction.channelType !== 'voice' &&
-          this.props.interaction.direction !== 'outbound' &&
+          this.props.interaction.direction !== 'agent-initiated' &&
           this.props.interaction.status !== 'wrapup' && (
           <div style={{ position: 'relative' }}>
             <Button